complete editor + image component
