+++ assets/src/components/editor/interfaces.ts
@@ -1,6 +1,5 @@
 import { ReactEditor } from 'slate-react'
 
-
 export type CommandDesc = {
   type: 'CommandDesc',
   icon: string,
@@ -12,3 +11,13 @@ export type Command = {
   precondition: (editor: ReactEditor) => void,
   execute: (editor: ReactEditor) => void,
 }
+
+// Width of padding on right hand side to allow toolbar toggler
+// to never obstruct text
+export const gutterWidth = 18;
+
+export type GroupDivider = {
+  type: 'GroupDivider'
+}
+
+export type ToolbarItem = CommandDesc | GroupDivider;
